# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "InKCre",
    "version": "0.1.0"
  },
  "paths": {
...
    "/crons": {
      "get": {
        "tags": [
          "cron"
        ],
        "summary": "List Crons",
        "operationId": "list_crons_crons_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "exclusiveMinimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response List Crons Crons Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "cron"
        ],
        "summary": "Create Cron",
        "operationId": "create_cron_crons_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CronForm"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CronModel"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/crons/{cron_id}": {
      "get": {
        "tags": [
          "cron"
        ],
        "summary": "Get Cron",
        "operationId": "get_cron_crons__cron_id__get",
        "parameters": [
          {
            "name": "cron_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Cron Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CronModel"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "cron"
        ],
        "summary": "Update Cron",
        "operationId": "update_cron_crons__cron_id__patch",
        "parameters": [
          {
            "name": "cron_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Cron Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CronUpdateForm"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CronModel"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "cron"
        ],
        "summary": "Delete Cron",
        "operationId": "delete_cron_crons__cron_id__delete",
        "parameters": [
          {
            "name": "cron_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Cron Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/crons/{cron_id}/run": {
      "post": {
        "tags": [
          "cron"
        ],
        "summary": "Run Cron",
        "operationId": "run_cron_crons__cron_id__run_post",
        "parameters": [
          {
            "name": "cron_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Cron Id"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobModel"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/agents": {
      "get": {
        "tags": [
          "agent"
        ],
        "summary": "List Agents",
        "operationId": "list_agents_agents_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "exclusiveMinimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response List Agents Agents Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "agent"
        ],
        "summary": "Create Agent",
        "operationId": "create_agent_agents_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentForm"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentDefinitionModel"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/agents/{agent_id}": {
      "get": {
        "tags": [
          "agent"
        ],
        "summary": "Get Agent",
        "operationId": "get_agent_agents__agent_id__get",
        "parameters": [
          {
            "name": "agent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Agent Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentDefinitionModel"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "agent"
        ],
        "summary": "Update Agent",
        "operationId": "update_agent_agents__agent_id__patch",
        "parameters": [
          {
            "name": "agent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Agent Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentUpdateForm"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentDefinitionModel"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "agent"
        ],
        "summary": "Delete Agent",
        "operationId": "delete_agent_agents__agent_id__delete",
        "parameters": [
          {
            "name": "agent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Agent Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/agent-tools": {
      "get": {
        "tags": [
          "agent"
        ],
        "summary": "List Agent Tools",
        "operationId": "list_agent_tools_agent_tools_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "exclusiveMinimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response List Agent Tools Agent Tools Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/agent-tools/{tool_id}": {
      "get": {
        "tags": [
          "agent"
        ],
        "summary": "Get Agent Tool",
        "operationId": "get_agent_tool_agent_tools__tool_id__get",
        "parameters": [
          {
            "name": "tool_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tool Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FunctionTool"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AgentDefinitionModel": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "system_prompt": {
            "type": "string",
            "title": "System Prompt"
          },
          "tools": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tools",
            "default": []
          },
          "tool_choice": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "none",
                  "auto",
                  "required"
                ]
              },
              {
                "$ref": "#/components/schemas/NamedToolChoice"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tool Choice"
          },
          "model": {
            "type": "integer",
            "title": "Model"
          },
          "max_model_calls_per_turn": {
            "type": "integer",
            "title": "Max Model Calls Per Turn"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "name",
          "system_prompt",
          "model",
          "max_model_calls_per_turn"
        ],
        "title": "AgentDefinitionModel",
        "description": "One reusable system-prompt, model, Tool and turn-budget composition."
      },
      "AgentForm": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "system_prompt": {
            "type": "string",
            "title": "System Prompt"
          },
          "model": {
            "type": "integer",
            "title": "Model"
          },
          "tools": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tools",
            "default": []
          },
          "tool_choice": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "none",
                  "auto",
                  "required"
                ]
              },
              {
                "$ref": "#/components/schemas/NamedToolChoice"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tool Choice"
          },
          "max_model_calls_per_turn": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Max Model Calls Per Turn"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "system_prompt",
          "model",
          "max_model_calls_per_turn"
        ],
        "title": "AgentForm",
        "description": "Editable Agent definition, without execution or database-managed fields."
      },
      "AgentUpdateForm": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "system_prompt": {
            "type": "string",
            "title": "System Prompt"
          },
          "model": {
            "type": "integer",
            "title": "Model"
          },
          "tools": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tools",
            "default": []
          },
          "tool_choice": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "none",
                  "auto",
                  "required"
                ]
              },
              {
                "$ref": "#/components/schemas/NamedToolChoice"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tool Choice"
          },
          "max_model_calls_per_turn": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Max Model Calls Per Turn"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "AgentUpdateForm",
        "description": "PATCH omission preserves a field; only tool_choice accepts explicit null."
      },
...
      "CronForm": {
        "properties": {
          "schedule": {
            "type": "string",
            "minLength": 1,
            "title": "Schedule"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "default": true
          },
          "job_type": {
            "type": "string",
            "title": "Job Type"
          },
          "job_parameters": {
            "additionalProperties": true,
            "type": "object",
            "title": "Job Parameters"
          },
          "job_timeout_seconds": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Timeout Seconds"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "schedule",
          "job_type"
        ],
        "title": "CronForm"
      },
      "CronModel": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "schedule": {
            "type": "string",
            "minLength": 1,
            "title": "Schedule"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "default": true
          },
          "job_type": {
            "type": "string",
            "title": "Job Type"
          },
          "job_parameters": {
            "additionalProperties": true,
            "type": "object",
            "title": "Job Parameters"
          },
          "job_timeout_seconds": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Timeout Seconds"
          },
          "last_job": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Job"
          },
          "last_scheduled_for": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Scheduled For"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "schedule",
          "job_type"
        ],
        "title": "CronModel",
        "description": "One five-field UNIX schedule that materializes typed Jobs."
      },
      "CronUpdateForm": {
        "properties": {
          "schedule": {
            "type": "string",
            "minLength": 1,
            "title": "Schedule"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled"
          },
          "job_type": {
            "type": "string",
            "title": "Job Type"
          },
          "job_parameters": {
            "additionalProperties": true,
            "type": "object",
            "title": "Job Parameters"
          },
          "job_timeout_seconds": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Timeout Seconds"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "CronUpdateForm",
        "description": "Optional submitted fields; null only clears the timeout override."
      },
...
      "FunctionTool": {
        "properties": {
          "id": {
            "type": "string",
            "minLength": 1,
            "title": "Id"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "input_schema": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonValue"
            },
            "type": "object",
            "title": "Input Schema"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "description",
          "input_schema"
        ],
        "title": "FunctionTool",
        "description": "One caller-owned function Tool exposed to an AI model."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JobModel": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "parameters": {
            "additionalProperties": true,
            "type": "object",
            "title": "Parameters"
          },
          "state": {
            "additionalProperties": true,
            "type": "object",
            "title": "State"
          },
          "abort_requested": {
            "type": "boolean",
            "title": "Abort Requested",
            "default": false
          },
          "timeout_seconds": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Timeout Seconds"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus",
            "default": "pending"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "closed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Closed At"
          }
        },
        "type": "object",
        "required": [
          "type",
          "timeout_seconds"
        ],
        "title": "JobModel",
        "description": "One durable execution opportunity for an exact registered command."
      },
      "JobStatus": {
        "type": "string",
        "enum": [
          "pending",
          "running",
          "finished",
          "failed",
          "timed_out",
          "aborted"
        ],
        "title": "JobStatus"
      },
...
      "JsonValue": {},
...
      "NamedToolChoice": {
        "properties": {
          "type": {
            "type": "string",
            "const": "function",
            "title": "Type",
            "default": "function"
          },
          "tool": {
            "type": "string",
            "minLength": 1,
            "title": "Tool"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "tool"
        ],
        "title": "NamedToolChoice",
        "description": "Require one exact caller-owned Tool."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}